deploy_bot.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
🤖 TELEGRAM BOT - DEPLOYMENT VERSION
Railway/Heroku için optimize edilmiş bot
"""

import asyncio
import logging
import os
import sys
from datetime import datetime

# Telegram bot imports
try:
    from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
    from telegram.ext import Application, CommandHandler, CallbackQueryHandler, ContextTypes
    print("✅ Telegram modülü yüklendi!")
except ImportError as e:
    print(f"❌ Telegram modülü yüklenemedi: {e}")
    sys.exit(1)

# Bot konfigürasyonu
BOT_TOKEN = os.environ.get('BOT_TOKEN', '8449268731:AAGWc5cFh2IxRnfxQsI_vtFLlWLLM8ZlIfc')
WEBAPP_URL = os.environ.get('WEBAPP_URL', 'https://localhost:5000')

# Logging ayarları
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Start komutu"""
    user = update.effective_user
    
    welcome_text = f"""
🔥 **CARD CHECKER PANEL BOT** 🔥

Merhaba {user.first_name}! 

Panel URL: {WEBAPP_URL}

🚀 **Komutlar:**
/panel - Web paneli aç
/stats - İstatistikler
/help - Yardım menüsü

⚡ **24/7 Kesintisiz Hizmet**
    """
    
    # Inline keyboard oluştur
    keyboard = [
        [InlineKeyboardButton("🌐 Panel Aç", url=WEBAPP_URL)],
        [InlineKeyboardButton("📊 İstatistikler", callback_data="stats")],
        [InlineKeyboardButton("🆘 Yardım", callback_data="help")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(
        welcome_text,
        reply_markup=reply_markup,
        parse_mode='Markdown'
    )
    
    print(f"✅ Kullanıcı bağlandı: {user.first_name} (@{user.username})")

async def panel_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Panel komutu"""
    keyboard = [
        [InlineKeyboardButton("🚀 Panel Aç", url=WEBAPP_URL)],
        [InlineKeyboardButton("📱 Mobil Görünüm", url=f"{WEBAPP_URL}/mobile")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    panel_text = f"""
🔥 **CARD CHECKER PANEL** 🔥

**Panel URL:** {WEBAPP_URL}

🌐 **Web Panel:** Tam özellikli panel
📱 **Mobil:** Optimize görünüm

⚡ **Özellikler:**
• Gerçek zamanlı kontrol
• Yüksek başarı oranı
• Güvenli bağlantı
• 24/7 hizmet
    """
    
    await update.message.reply_text(
        panel_text,
        reply_markup=reply_markup,
        parse_mode='Markdown'
    )

async def stats_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """İstatistik komutu"""
    import random
    
    stats_text = f"""
📊 **PANEL İSTATİSTİKLERİ** 📊

🔢 **Toplam:** {random.randint(1000, 5000):,}
✅ **Canlı:** {random.randint(100, 500):,}
❌ **Ölü:** {random.randint(500, 2000):,}
📈 **Başarı:** {random.uniform(15.0, 25.0):.1f}%
⚡ **Hız:** {random.randint(80, 150)} CPM
⏱️ **Uptime:** 24/7

🕐 **Güncelleme:** {datetime.now().strftime('%H:%M:%S')}
🤖 **Bot:** ✅ Online
🌐 **Panel:** ✅ Aktif
    """
    
    keyboard = [
        [InlineKeyboardButton("🔄 Yenile", callback_data="stats")],
        [InlineKeyboardButton("🌐 Panel", url=WEBAPP_URL)]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(
        stats_text,
        reply_markup=reply_markup,
        parse_mode='Markdown'
    )

async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Yardım komutu"""
    help_text = """
🆘 **YARDIM MENÜSÜ** 🆘

📋 **Komutlar:**
/start - Bot'u başlat
/panel - Web paneli aç
/stats - İstatistikler
/help - Bu yardım menüsü

🌐 **Panel Özellikleri:**
• Gerçek zamanlı kart kontrolü
• Canlı istatistikler
• Modern arayüz
• Mobil uyumlu
• SSL güvenliği

🔒 **Güvenlik:**
• Şifreli bağlantı
• Yetkili erişim
• Rate limiting
• 24/7 monitoring

📞 **Destek:**
Bu bot 24/7 aktif olarak çalışmaktadır.
Panel ile ilgili sorunlarınız için admin ile iletişime geçin.

⚡ **Kesintisiz Hizmet**
    """
    
    keyboard = [
        [InlineKeyboardButton("🌐 Panel Aç", url=WEBAPP_URL)],
        [InlineKeyboardButton("📊 İstatistikler", callback_data="stats")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(
        help_text,
        reply_markup=reply_markup,
        parse_mode='Markdown'
    )

async def button_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Button callback"""
    query = update.callback_query
    await query.answer()
    
    if query.data == "stats":
        await stats_command(update, context)
    elif query.data == "help":
        await help_command(update, context)
    elif query.data == "panel":
        keyboard = [[InlineKeyboardButton("🚀 Panel Aç", url=WEBAPP_URL)]]
        reply_markup = InlineKeyboardMarkup(keyboard)
        await query.edit_message_text(
            f"🔥 **PANEL ERİŞİMİ** 🔥\n\nPanel: {WEBAPP_URL}",
            reply_markup=reply_markup,
            parse_mode='Markdown'
        )

def main():
    """Ana fonksiyon"""
    print("🤖 Telegram Bot başlatılıyor...")
    print(f"🔑 Token: {BOT_TOKEN[:10]}...")
    print(f"🌐 Panel URL: {WEBAPP_URL}")
    print("=" * 50)
    
    try:
        # Application oluştur
        application = Application.builder().token(BOT_TOKEN).build()
        
        # Handler'ları ekle
        application.add_handler(CommandHandler("start", start_command))
        application.add_handler(CommandHandler("panel", panel_command))
        application.add_handler(CommandHandler("stats", stats_command))
        application.add_handler(CommandHandler("help", help_command))
        application.add_handler(CallbackQueryHandler(button_callback))
        
        # Bot'u çalıştır
        print("🚀 Bot başlatıldı! 24/7 aktif.")
        application.run_polling(allowed_updates=Update.ALL_TYPES)
        
    except Exception as e:
        print(f"❌ Hata: {e}")
        logger.error(f"Bot hatası: {e}")

if __name__ == '__main__':
    try:
        main()
    except KeyboardInterrupt:
        print("\n🛑 Bot kapatılıyor...")
    except Exception as e:
        print(f"❌ Hata: {e}")
        logger.error(f"Ana hata: {e}")