deploy_requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-telegram-bot==22.5
python-dotenv==1.1.1
psutil==5.9.6
requests==2.31.0
Werkzeug==2.3.7
eventlet==0.33.3